refactor(goal-detail): split the pure model logic out of goalDetailShared (#467)
First slice of the goalDetailShared breakup (keeping it from becoming a new shared
god file). Move the pure logic — calcDeadlineMonths + computeGoalCalculator (the
goal projection) and describeHistoryRow (the transaction-history row descriptor) —
into a new `goalDetailModel.ts`, leaving goalDetailShared for its presentational
UI. The three consumers (GoalDetailSheet, DesktopGoalDetail, the shared test) now
import these from goalDetailModel; the now-unused lucide icons + txKind/TxKind*
imports drop out of goalDetailShared. No behavior change.

Verified: full unit suite 1261 passed (the goalDetailShared test exercises the
calculator + history-row descriptor from the new module); typecheck 0 errors;
lint 0 errors; git diff --check clean. goalDetailShared 717 → 647 lines.
